re PR fortran/32770 ([Meta-bug] -fdefault-integer-8 issues)
2007-11-20  Thomas Koenig  <tkoenig@gcc.gnu.org>

	PR fortran/32770
	* gfortran.dg/record_marker_1.f90:  Explicitly set kinds
	so test passes with -frecord-marker=8.
	* gfortran.dg/integer_exponentiation_1.f90:  Likewise.
	* gfortran.dg/integer_exponentiation_2.f90:  Likewise.
	* gfortran.dg/integer_exponentiation_3.F90:  Likewise.
	* gfortran.dg/integer_exponentiation_5.F90:  Likewise.
	* gfortrand.dg/modulo_1.f90:  Likewise.

From-SVN: r130322

+++ gcc/testsuite/gfortran.dg/modulo_1.f90
@@ -1,6 +1,6 @@
 ! { dg-do compile }
 ! PR fortran/23912
-  integer i4
+  integer(kind=4) i4
   integer(kind=8) i8
 
   i4 = modulo(i4,i8) ! { dg-warning "Extension" }
